force enthalpy increment as arg in gurvich h, g
Turns out that calculating (H(298.15) - H(0)) with dh0_gurvich
fails when the polynomial in question doesn't work at room temperature,
resulting in erroneous piecewise Gibbs and enthalpy functions for
high-temperature pieces in the piecewise functions.

The solution is to force this enthalpy increment as an argument,
exactly as in the first versions of gurvich functions.